- Bug fix in the launch script (it didn't overwrite the existing binary for Scala binaries) - Scala documentation - Implementation of K3.Filter in the M3-to-K3 translation phase - Optimization for nested filters
git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@2586 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 6b65be97937cccbcc992ab028840919ea857e96b

+++ lib/dbt_scala/src/org/dbtoaster/dbtoasterlib/ImplicitConversions.scala
@@ -4,6 +4,6 @@ package org.dbtoaster.dbtoasterlib {
     * used by queries.
     */
   object ImplicitConversions {
-    implicit def boolToLong(b: Boolean): Long = if(b) 1.toLong else 0.toLong 
+    implicit def boolToLong(b: Boolean): Long = if(b) 1L else 0L
   }
 }
